refactor ccp4 header module
