+++ pyworkflow/em/packages/ccp4/protocol_refmac.py
@@ -26,18 +26,19 @@
 
 import os
 import stat
+
 import pyworkflow.protocol.constants as const
+from constants import CCP4VERSION, CCP4VERSIONFILENAME
 from pyworkflow import VERSION_1_2
 from pyworkflow.em import PdbFile
+from pyworkflow.em.headers.ccp4header import (
+    adaptFileToCCP4, START, Ccp4Header)
+from pyworkflow.em.packages.ccp4.convert import (runCCP4Program, getProgram)
 from pyworkflow.em.packages.ccp4.refmac_template_map2mtz import \
     template_map_mtz
 from pyworkflow.em.packages.ccp4.refmac_template_refine import template_refine
 from pyworkflow.em.protocol import EMProtocol
-from pyworkflow.em.header_handler.CCP4.convert import (
-    adaptFileToCCP4, START, Ccp4Header)
-from pyworkflow.em.packages.ccp4.convert import (runCCP4Program, getProgram)
 from pyworkflow.protocol.params import PointerParam, IntParam, FloatParam
-from constants import CCP4VERSION, CCP4VERSIONFILENAME
 
 
 class CCP4ProtRunRefmac(EMProtocol):
