+++ pyworkflow/em/packages/xmipp3/tests/test_protocol_extract_unit_cell.py
@@ -27,17 +27,16 @@
 # that are available to select in the form for the extraction of the unit
 # cell from a specific volume
 
-import os
 import math
+import os
 from tempfile import mkstemp
 
-from pyworkflow.em.symmetry import Icosahedron
 from pyworkflow.em.constants import SYM_I222r, SYM_I222, SCIPION_SYM_NAME, \
     SYM_In25, SYM_In25r, SYM_CYCLIC, SYM_DIHEDRAL, SYM_TETRAHEDRAL, \
     SYM_OCTAHEDRAL
 from pyworkflow.em.convert import ImageHandler
 from pyworkflow.em.data import Transform
-from pyworkflow.em.header_handler.CCP4.convert import Ccp4Header
+from pyworkflow.em.headers.ccp4header import Ccp4Header
 from pyworkflow.em.packages.xmipp3 import getEnviron
 from pyworkflow.em.packages.xmipp3.constants import XMIPP_SYM_NAME
 from pyworkflow.em.packages.xmipp3.pdb.protocol_pseudoatoms \
@@ -47,6 +46,7 @@ from pyworkflow.em.packages.xmipp3.pdb.protocol_pseudoatoms_base \
 from pyworkflow.em.packages.xmipp3.protocol_extract_unit_cell \
     import XmippProtExtractUnit
 from pyworkflow.em.protocol import ProtImportVolumes
+from pyworkflow.em.symmetry import Icosahedron
 from pyworkflow.tests import BaseTest, setupTestProject
 from pyworkflow.utils import runJob
