+++ pyworkflow/em/protocol/protocol_import/volumes.py
@@ -30,18 +30,16 @@ Particles, Volumes...
 """
 
 from os.path import exists, basename, abspath
-from pyworkflow.utils.properties import Message
-from pyworkflow.utils.path import createAbsLink, copyFile
-import pyworkflow.protocol.constants as const
+
 import pyworkflow.protocol.params as params
+from base import ProtImportFiles
+from images import ProtImportImages
 from pyworkflow.em import Volume, ImageHandler, PdbFile
 from pyworkflow.em.convert import downloadPdb
 from pyworkflow.em.data import Transform
-from base import ProtImportFiles
-from images import ProtImportImages
-from pyworkflow.em.header_handler.CCP4.convert import Ccp4Header, \
-    adaptFileToCCP4, ORIGIN
-
+from pyworkflow.em.headers.ccp4header import adaptFileToCCP4, ORIGIN
+from pyworkflow.utils.path import createAbsLink, copyFile
+from pyworkflow.utils.properties import Message
 
 
 class ProtImportVolumes(ProtImportImages):
