+++ pyworkflow/em/wizard.py
@@ -28,37 +28,33 @@
 This module implement some wizards
 """
 
-import os
-from os.path import basename, exists
 import Tkinter as tk
+import os
 import ttk
 
-from pyworkflow.wizard import Wizard
-import pyworkflow.gui.dialog as dialog
-from pyworkflow.gui.widgets import LabelSlider
-from pyworkflow.gui.tree import BoundTree, TreeProvider
-from pyworkflow import findResource
-from pyworkflow.object import PointerList, Pointer
+import xmipp
 
-from pyworkflow.em.convert import ImageHandler
-from pyworkflow.em.constants import (UNIT_PIXEL, 
+import pyworkflow.gui.dialog as dialog
+from pyworkflow import findResource
+from pyworkflow.em.constants import (UNIT_PIXEL,
                                      UNIT_PIXEL_FOURIER,
                                      UNIT_ANGSTROM,
-                                     UNIT_ANGSTROM_FOURIER,
-                                     FILTER_LOW_PASS, 
-                                     FILTER_BAND_PASS, 
+                                     FILTER_LOW_PASS,
+                                     FILTER_BAND_PASS,
                                      FILTER_HIGH_PASS
                                      )
+from pyworkflow.em.convert import ImageHandler
 from pyworkflow.em.data import (Volume, SetOfMicrographs, SetOfParticles,
                                 SetOfVolumes)
+from pyworkflow.em.headers.ccp4header import Ccp4Header
 from pyworkflow.em.protocol.protocol_import import (ProtImportImages,
-                                                    ProtImportMovies,
                                                     ProtImportCoordinates,
                                                     ProtImportVolumes)
+from pyworkflow.gui.tree import BoundTree, TreeProvider
+from pyworkflow.gui.widgets import LabelSlider
+from pyworkflow.object import PointerList, Pointer
+from pyworkflow.wizard import Wizard
 
-
-import xmipp
-from pyworkflow.em.header_handler.CCP4.convert import Ccp4Header
 
 #===============================================================================
 #    Wizard EM base class
